Refactor sessionstateitem into seperate file
